+++ src/common/Environment.ts
@@ -102,12 +102,24 @@ args
   .option(
     ["d", "debug"],
     "Optional. Enable debug log by providing a valid local file path as log destination"
+  )
+  .option(
+    ["", "disableProductStyleUrl"],
+    "Optional. Disable getting account name from the host of request Uri, always get account name from the first path segment of request Uri"
+  )
+  .option(
+    ["", "disableTelemetry"],
+    "Optional. Disable telemtry collection of Azurite. If not specify this parameter Azurite will collect telemetry data by default."
   );
 
 (args as any).config.name = "azurite";
 
 export default class Environment implements IEnvironment {
   private flags = args.parse(process.argv);
+
+  //public ToString(): string | undefined {
+  //  return Object.getOwnPropertyNames(this.flags).join(", ");
+  //}
 
   public blobHost(): string | undefined {
     return this.flags.blobHost;
@@ -212,6 +224,14 @@ export default class Environment implements IEnvironment {
 
   public extentMemoryLimit(): number | undefined {
     return this.flags.extentMemoryLimit;
+  }
+
+  public disableTelemetry(): boolean {
+    if (this.flags.disableTelemetry !== undefined) {
+      return true;
+    }
+    // default is false which will collect telemetry data
+    return false;
   }
 
   public async debug(): Promise<string | undefined> {