+++ src/common/VSCEnvironment.ts
@@ -5,7 +5,7 @@ import { window, workspace, WorkspaceFolder } from "vscode";
 import IEnvironment from "./IEnvironment";
 
 export default class VSCEnvironment implements IEnvironment {
-  private workspaceConfiguration = workspace.getConfiguration("azurite");
+  public workspaceConfiguration = workspace.getConfiguration("azurite");
 
   public blobHost(): string | undefined {
     return this.workspaceConfiguration.get<string>("blobHost");
@@ -128,5 +128,11 @@ export default class VSCEnvironment implements IEnvironment {
 
   public extentMemoryLimit(): number | undefined {
     return this.workspaceConfiguration.get<number>("extentMemoryLimit");
+  }
+
+  public disableTelemetry(): boolean {
+    return (
+      this.workspaceConfiguration.get<boolean>("disableTelemetry") || false
+    );
   }
 }