+++ src/table/main.ts
@@ -7,6 +7,7 @@ import TableConfiguration from "./TableConfiguration";
 import TableEnvironment from "./TableEnvironment";
 import TableServer from "./TableServer";
 import { DEFAULT_TABLE_LOKI_DB_PATH } from "./utils/constants";
+import { AzuriteTelemetryClient } from "../common/Telemetry";
 
 // tslint:disable:no-console
 
@@ -67,12 +68,16 @@ async function main() {
   console.log(beforeStartMessage);
   await server.start();
   console.log(afterStartMessage);
+  
+  AzuriteTelemetryClient.init(location, !env.disableTelemetry(), env);
+  await AzuriteTelemetryClient.TraceStartEvent("Table");
 
   // Handle close event
   process
     .once("message", (msg) => {
       if (msg === "shutdown") {
         console.log(beforeCloseMessage);
+        AzuriteTelemetryClient.TraceStopEvent("Table");
         server.close().then(() => {
           console.log(afterCloseMessage);
         });
@@ -80,6 +85,7 @@ async function main() {
     })
     .once("SIGINT", () => {
       console.log(beforeCloseMessage);
+      AzuriteTelemetryClient.TraceStopEvent("Table");
       server.close().then(() => {
         console.log(afterCloseMessage);
       });